patch.sh allows live patching of game from git repository

+++ Server/patch.sh
@@ -12,6 +12,7 @@ else
    echo "Abortng by user request."
    exit 1
 fi
+type=0
 if [ -f ./src.tbz ]
 then
    echo "I found a source file."
@@ -26,16 +27,29 @@ then
       exit 1
    fi
 else
-   echo "I'm sorry, I did not find a src.tbz at '$(pwd)'.  Please copy that file in and try again."
-   exit 1
+   echo "Hum.  No source files.  I'll tell git to yoink the source files for you then."
+   echo "downloading..."|tr -d '\012'
+   git clone https://github.com/RhostMUSH/trunk rhost_tmp > /dev/null 2>&1
+   echo "done!"
+   type=1
 fi
 echo "Making a backup of all your files, please wait..."|tr -d '\012'
 lc_date=$(date +%m%d%y%H%M%S)
 tar -czf src_backup_${lc_date}.tgz src/*.c hdrs/*.h game/txt/help.txt game/txt/wizhelp.txt > /dev/null 2>&1
 echo "... completed.  Filename is src_backup_${lc_date}.tgz"
-echo "Copying your binary ... just in case.  Backup will be src/netrhost.automate"
-cp -f src/netrhost src/netrhost.automate
-bunzip -cd src.tbz|tar -xvf -
+echo "Copying your binary ... just in case.  Backup will be src/netrhost.automate (or bin/netrhost.automate)"
+[[ -f src/netrhost ]] && cp -f src/netrhost src/netrhost.automate
+[[ -f bin/netrhost ]] && cp -f bin/netrhost bin/netrhost.automate
+if [ ${type} -eq 0 ]
+then
+   bunzip -cd src.tbz|tar -xvf -
+else
+   cp rhost_tmp/Server/src/*.c src
+   cp rhost_tmp/Server/hdrs/*.h hdrs
+   cp rhost_tmp/Server/game/txt/help.txt game/txt/help.txt
+   cp rhost_tmp/Server/game/txt/wizhelp.txt game/txt/wizhelp.txt
+   rm -rf ./rhost_tmp
+fi
 cd src
 make clean;make
 cd ../game/txt
